v14
kritik sistem güncellemesi

+++ admin/functions/category-delete.php
@@ -3,6 +3,12 @@
 // Ana fonskiyon dosyası ekleniyor.
 include("../../codex.php");
 
+// Güvenlik kontrolü yapılıyor.
+if (!isset($_SESSION["user_role"]) and $_SESSION["user_role"] != "admin") {
+    header("Location:" . $site_name . "/?status=permission-exist");
+    exit();
+}
+
 $category_id = $_GET["category_id"];
 
 $categories = $db->prepare("DELETE from categories where category_id=:id");
